+++ tests/unit/specs/views/Apps.spec.js
@@ -1,12 +1,12 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils';
-import Apps from '@/views/Apps.vue';
+import Apps from '@/views/Apps/index.vue';
 import VueRouter from 'vue-router';
 const router = new VueRouter();
 
 const localVue = createLocalVue();
 localVue.use(VueRouter);
 
-describe('App.vue', () => {
+describe('App/index.vue', () => {
   let wrapper;
   beforeEach(() => {
     wrapper = shallowMount(Apps, {