+++ tests/unit/specs/views/Discovery.spec.js
@@ -7,8 +7,8 @@ jest.mock('@weni/unnnic-system', () => ({
 
 import Vuex from 'vuex';
 import { mount, createLocalVue } from '@vue/test-utils';
-import Discovery from '@/views/Discovery.vue';
-import AppGrid from '@/components/AppGrid.vue';
+import Discovery from '@/views/Discovery/index.vue';
+import AppGrid from '@/components/AppGrid/index.vue';
 import { singleApp } from '../../../__mocks__/appMock';
 
 const genericApp = {
@@ -56,12 +56,27 @@ const mountComponent = async ({ createAppCode = null, apps = [singleApp] } = {})
     externalServicesList: apps,
   };
 
+  const ecommerceAppsActions = {
+    getEcommerceTypes: jest.fn(),
+  };
+
+  const ecommerceAppsState = {
+    loadingEcommerceApps: false,
+    errorEcommerceApps: null,
+    ecommerceAppsList: apps,
+  };
+
   const store = new Vuex.Store({
     modules: {
       externals: {
         namespaced: true,
         actions: externalServicesActions,
         state: externalServicesState,
+      },
+      ecommerce: {
+        namespaced: true,
+        actions: ecommerceAppsActions,
+        state: ecommerceAppsState,
       },
     },
     actions,
@@ -91,7 +106,7 @@ const mountComponent = async ({ createAppCode = null, apps = [singleApp] } = {})
 const localVue = createLocalVue();
 localVue.use(Vuex);
 
-describe('Discovery.vue', () => {
+describe('Discovery/index.vue', () => {
   beforeEach(() => {
     jest.clearAllMocks();
     jest.restoreAllMocks();
@@ -149,19 +164,23 @@ describe('Discovery.vue', () => {
       const { wrapper, store } = await mountComponent();
       store.state.appType.allAppTypes = [singleApp, genericApp];
       store.state.externals.externalServicesList = [singleApp, genericApp];
+      store.state.ecommerce.ecommerceAppsList = [singleApp, genericApp];
 
       expect(wrapper.vm.filteredApps).toEqual([singleApp, genericApp]);
       expect(wrapper.vm.filteredExternalServices).toEqual([singleApp, genericApp]);
+      expect(wrapper.vm.filteredEcommerceApps).toEqual([singleApp, genericApp]);
       expect(wrapper.vm.filteredBiApps).toEqual(wrapper.vm.biApps);
 
       wrapper.vm.searchTerm = 'weni';
       expect(wrapper.vm.filteredApps).toEqual([singleApp]);
       expect(wrapper.vm.filteredExternalServices).toEqual([singleApp]);
+      expect(wrapper.vm.filteredEcommerceApps).toEqual([singleApp]);
       expect(wrapper.vm.filteredBiApps).toEqual([]);
 
       wrapper.vm.searchTerm = ' ';
       expect(wrapper.vm.filteredApps).toEqual([singleApp, genericApp]);
       expect(wrapper.vm.filteredExternalServices).toEqual([singleApp, genericApp]);
+      expect(wrapper.vm.filteredEcommerceApps).toEqual([singleApp, genericApp]);
       expect(wrapper.vm.filteredBiApps).toEqual(wrapper.vm.biApps);
     });
 
@@ -169,9 +188,27 @@ describe('Discovery.vue', () => {
       const { wrapper, store } = await mountComponent();
       store.state.appType.allAppTypes = null;
       store.state.externals.externalServicesList = null;
+      store.state.ecommerce.ecommerceAppsList = null;
 
       expect(wrapper.vm.filteredApps).toEqual([]);
       expect(wrapper.vm.filteredExternalServices).toEqual([]);
+      expect(wrapper.vm.filteredEcommerceApps).toEqual([]);
+    });
+  });
+
+  describe('hasAnyVisibleApp', () => {
+    it('should return true if there are any visible app', async () => {
+      const { wrapper } = await mountComponent();
+      wrapper.vm.searchTerm = 'weni';
+
+      expect(wrapper.vm.hasAnyVisibleApp).toBeTruthy();
+    });
+
+    it('should return false if there are no visible app', async () => {
+      const { wrapper } = await mountComponent();
+      wrapper.vm.searchTerm = 'unexisting app';
+
+      expect(wrapper.vm.hasAnyVisibleApp).toBeFalsy();
     });
   });
 });